+++ back-end/src/main/java/com/tdd/backend/option/model/Category.java
@@ -1,4 +1,4 @@
-package com.tdd.backend.option;
+package com.tdd.backend.option.model;
 
 import java.util.Arrays;
