+++ back-end/src/main/java/com/tdd/backend/option/model/RenderOption.java
@@ -1,4 +1,4 @@
-package com.tdd.backend.option.render;
+package com.tdd.backend.option.model;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Table;
@@ -8,10 +8,10 @@ import lombok.Getter;
 @Table("entire_options")
 @Getter
 public class RenderOption {
-	@Id
-	private Long id;
 	private final String optionName;
 	private final String category;
+	@Id
+	private Long id;
 
 	public RenderOption(String optionName, String category) {
 		this.optionName = optionName;
