+++ back-end/src/main/java/com/tdd/backend/option/repository/RenderRepository.java
@@ -1,4 +1,4 @@
-package com.tdd.backend.option.render;
+package com.tdd.backend.option.repository;
 
 import java.util.List;
 
@@ -8,6 +8,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
+
+import com.tdd.backend.option.model.RenderOption;
 
 @Repository
 public class RenderRepository {
