+++ back-end/src/main/java/com/tdd/backend/post/model/Option.java
@@ -3,8 +3,8 @@ package com.tdd.backend.post.model;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Table;
 
-import com.tdd.backend.option.Category;
 import com.tdd.backend.option.data.OptionDto;
+import com.tdd.backend.option.model.Category;
 
 @Table("options")
 public class Option {
